+++ projects/gleks/ui/src/public-api.ts
@@ -37,6 +37,7 @@ export * from './lib/services/theme-service/theme.service';
 export * from './lib/shared/types';
 export * from './lib/shared/config';
 export * from './lib/shared/dropdown-base';
+export * from './lib/shared/float-label-state';
 export type { GogDropdownDirection } from './lib/shared/dropdown-position';
 export type { GogTooltipSide } from './lib/shared/tooltip-position';
 export type { GogIconName } from './lib/shared/icons';